refactor: split armor info to another file
