+++ src/iteminfo_format_utils.cpp
@@ -1,0 +1,8 @@
+#include "iteminfo_format_utils.h"
+
+void insert_separation_line( std::vector<iteminfo> &info )
+{
+    if( info.empty() || info.back().sName != "--" ) {
+        info.push_back( iteminfo( "DESCRIPTION", "--" ) );
+    }
+}
